apply feedback

+++ tools/pin_github_action.py
@@ -1,13 +1,8 @@
 from __future__ import annotations
 
-import os
-
-GITHUB_TOKEN = os.environ.get("GITHUB_TOKEN")
-if not GITHUB_TOKEN:
-    raise SystemExit("GITHUB_TOKEN not set.")
-
 import argparse
 import json
+import os
 import re
 import urllib.request
 from functools import lru_cache
@@ -16,7 +11,7 @@ from urllib.error import HTTPError
 
 
 @lru_cache(maxsize=None)
-def get_sha(repo: str, ref: str) -> str:
+def get_sha(repo: str, ref: str, github_token: str) -> str:
     if len(ref) == 40:
         try:
             int(ref, 16)
@@ -31,7 +26,7 @@ def get_sha(repo: str, ref: str) -> str:
                 method="GET",
                 headers={
                     "Accept": "application/vnd.github+json",
-                    "Authorization": f"token {GITHUB_TOKEN}",
+                    "Authorization": f"token {github_token}",
                     # A user agent is required. Lol.
                     "User-Agent": "python-requests/2.26.0",
                 },
@@ -42,12 +37,9 @@ def get_sha(repo: str, ref: str) -> str:
         if e.code == 403:
             print("You most likely didn't authorize your token for SAML to the getsentry org.")
         return ref
-    if resp.status != 200:
-        print(f"Status {resp.status} while resolving {ref} for {repo}.")
-        return ref
 
-    data = json.load(resp)
-    return data["sha"]  # type: ignore
+    data: dict[str, str] = json.load(resp)
+    return data["sha"]
 
 
 def extract_repo(action: str) -> str:
@@ -60,26 +52,30 @@ def extract_repo(action: str) -> str:
 
 def main(argv: Sequence[str] | None = None) -> int:
     parser = argparse.ArgumentParser()
-    parser.add_argument("file", nargs="+", type=argparse.FileType("r+"), help="github actions file")
+    parser.add_argument("files", nargs="+", type=str, help="path to github actions file")
     args = parser.parse_args(argv)
-    files = args.file
+
+    GITHUB_TOKEN = os.environ.get("GITHUB_TOKEN")
+    if not GITHUB_TOKEN:
+        raise SystemExit("GITHUB_TOKEN not set.")
 
     ACTION_VERSION_RE = re.compile(r"(?<=uses: )(?P<action>.*)@(?P<ref>.+?)\b")
-    for f in files:
-        newlines = []
-        for line in f:
-            if "uses:" in line and "@" in line:
+    for fp in args.files:
+        with open(fp, "r+") as f:
+            newlines = []
+            for line in f:
                 m = ACTION_VERSION_RE.search(line)
                 if not m:
+                    newlines.append(line)
                     continue
                 d = m.groupdict()
-                sha = get_sha(extract_repo(d["action"]), ref=d["ref"])
+                sha = get_sha(extract_repo(d["action"]), ref=d["ref"], github_token=GITHUB_TOKEN)
                 if sha != d["ref"]:
                     line = ACTION_VERSION_RE.sub(rf"\1@{sha}  # \2", line)
-            newlines.append(line)
-        f.seek(0)
-        f.truncate()
-        f.writelines(newlines)
+                newlines.append(line)
+            f.seek(0)
+            f.truncate()
+            f.writelines(newlines)
 
     return 0
